+++ src/constants.js
@@ -78,6 +78,7 @@ export const DEFAULT_HEIGHT = 'auto';
 export const DEFAULT_GAMMA = 1;
 
 // Default styles
+export const MIN_POINT_SIZE = 1;
 export const DEFAULT_POINT_SIZE = 6;
 export const DEFAULT_POINT_SIZE_SELECTED = 2;
 export const DEFAULT_POINT_OUTLINE_WIDTH = 2;